+++ spec/graphql/schema_spec.rb
@@ -479,11 +479,17 @@ To add other types to your schema, you might want `extra_types`: https://graphql
 
   describe "DidYouMean support" do
     class DidYouMeanSchema < GraphQL::Schema
+      class ExampleEnum < GraphQL::Schema::Enum
+        value "VALUE_ONE", "The first value"
+        value "VALUE_TWO", "The second value"
+      end
+
       class Query < GraphQL::Schema::Object
         field :first_field, String
         field :second_field, String
         field :second_fiel, String
         field :scond_field, String
+        field :third_field, ExampleEnum
       end
 
       query(Query)
@@ -507,8 +513,16 @@ To add other types to your schema, you might want `extra_types`: https://graphql
       res = no_dym_schema.execute("{ seconField }")
       assert_equal ["Field 'seconField' doesn't exist on type 'Query'"], res["errors"].map { |err| err["message"] }
     end
-  end 
-  
+
+    it "returns helpful message when non existing field is queried on a non-fields type" do
+      res = DidYouMeanSchema.execute("{ thirdField { foo } }")
+      assert_equal ["Selections can't be made on enums (field 'thirdField' returns ExampleEnum but has selections [\"foo\"])"], res["errors"].map { |err| err["message"] }
+
+      res = DidYouMeanSchema.execute("{ secondField { foo } }")
+      assert_equal ["Selections can't be made on scalars (field 'secondField' returns String but has selections [\"foo\"])"], res["errors"].map { |err| err["message"] }
+    end
+  end
+
   it "defers root type blocks until those types are used" do
     calls = []
     schema = Class.new(GraphQL::Schema) do